requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
telethon==1.28.5
python-telegram-bot==20.7
supabase==2.3.2
httpx==0.25.2
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==1.10.13
aiohttp==3.9.1